# tsc(perf): avoid large unions