--- a/static/app/components/activity/item/bubble.tsx
+++ b/static/app/components/activity/item/bubble.tsx
@@ -1,15 +1,18 @@
-import styled from '@emotion/styled';
+import styled, {StyledComponent} from '@emotion/styled';
 
-type Props = {
+interface ActivityBubbleProps {
   backgroundColor?: string;
   borderColor?: string;
-};
+}
 
 /**
  * This creates a bordered box that has a left pointing arrow
  * on the left-side at the top.
  */
-const ActivityBubble = styled('div')<Props>`
+
+const ActivityBubble: StyledComponent<ActivityBubbleProps> = styled(
+  'div'
+)<ActivityBubbleProps>`
   flex: 1;
   background-color: ${p => p.backgroundColor || p.theme.background};
   border: 1px solid ${p => p.borderColor || p.theme.border};