--- a/static/app/components/activity/note/input.tsx
+++ b/static/app/components/activity/note/input.tsx
@@ -3,7 +3,7 @@ import {Mention, MentionsInput, MentionsInputProps} from 'react-mentions';
 import {withTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 
-import Button from 'sentry/components/button';
+import Button, {ButtonProps} from 'sentry/components/button';
 import NavTabs from 'sentry/components/navTabs';
 import {IconMarkdown} from 'sentry/icons';
 import {t} from 'sentry/locale';
@@ -365,7 +365,7 @@ const Footer = styled('div')`
 
 type FooterButtonProps = {
   error?: string | null;
-} & React.ComponentProps<typeof Button>;
+} & ButtonProps;
 
 const FooterButton = styled(Button)<FooterButtonProps>`
   font-size: 13px;