--- a/static/app/components/buttonBar.tsx
+++ b/static/app/components/buttonBar.tsx
@@ -3,14 +3,14 @@ import {css} from '@emotion/react';
 import styled from '@emotion/styled';
 import classNames from 'classnames';
 
-import Button, {StyledButton} from 'sentry/components/button';
+import {ButtonProps, StyledButton} from 'sentry/components/button';
 import space, {ValidSize} from 'sentry/styles/space';
 
 type ButtonBarProps = {
   children: React.ReactNode;
   gap?: ValidSize;
   merged?: boolean;
-  active?: React.ComponentProps<typeof Button>['barId'];
+  active?: ButtonProps['barId'];
   className?: string;
 };