--- a/static/app/components/charts/transparentLoadingMask.tsx
+++ b/static/app/components/charts/transparentLoadingMask.tsx
@@ -3,22 +3,20 @@ import styled from '@emotion/styled';
 
 import LoadingMask from 'sentry/components/loadingMask';
 
-type Props = {
+interface Props {
   visible: boolean;
   className?: string;
   children?: React.ReactNode;
-} & React.ComponentProps<typeof LoadingMask>;
+}
 
-const TransparentLoadingMask = styled(
-  ({className, visible, children, ...props}: Props) => {
-    const other = visible ? {...props, 'data-test-id': 'loading-placeholder'} : props;
-    return (
-      <LoadingMask className={className} {...other}>
-        {children}
-      </LoadingMask>
-    );
-  }
-)<Props>`
+const TransparentLoadingMask = styled(({className, visible, children}: Props) => {
+  const other = visible ? {'data-test-id': 'loading-placeholder'} : {};
+  return (
+    <LoadingMask className={className} {...other}>
+      {children}
+    </LoadingMask>
+  );
+})<Props>`
   ${p => !p.visible && 'display: none;'};
   opacity: 0.4;
   z-index: 1;