--- a/static/app/components/navigationButtonGroup.tsx
+++ b/static/app/components/navigationButtonGroup.tsx
@@ -1,6 +1,6 @@
 import {LocationDescriptor} from 'history';
 
-import Button from 'sentry/components/button';
+import Button, {ButtonProps} from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
 import {IconNext, IconPrevious} from 'sentry/icons';
 import {t} from 'sentry/locale';
@@ -14,7 +14,7 @@ type Props = {
   hasNext?: boolean;
   hasPrevious?: boolean;
   className?: string;
-  size?: React.ComponentProps<typeof Button>['size'];
+  size?: ButtonProps['size'];
   onOldestClick?: () => void;
   onOlderClick?: () => void;
   onNewerClick?: () => void;