--- a/static/app/icons/iconAnchor.tsx
+++ b/static/app/icons/iconAnchor.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 
-import SvgIcon from './svgIcon';
+import SvgIcon, {SVGIconProps} from './svgIcon';
 
-type Props = React.ComponentProps<typeof SvgIcon>;
-
-const IconAnchor = React.forwardRef(function IconAnchor(
-  props: Props,
+const IconAnchor = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconAnchor(
+  props: SVGIconProps,
   ref: React.Ref<SVGSVGElement>
 ) {
   return (