--- a/static/app/icons/iconBusiness.tsx
+++ b/static/app/icons/iconBusiness.tsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 
 import {uniqueId} from 'sentry/utils/guid';
 
-import SvgIcon from './svgIcon';
+import SvgIcon, {SVGIconProps} from './svgIcon';
 
 type WrappedProps = {
   forwardRef: React.Ref<SVGSVGElement>;
@@ -56,7 +56,7 @@ function IconBusinessComponent({
   );
 }
 
-type Props = {
+interface Props extends SVGIconProps {
   /**
    * Renders a pink purple gradient on the icon
    */
@@ -66,11 +66,13 @@ type Props = {
    * Adds an animated shine to the icon
    */
   withShine?: boolean;
-} & React.ComponentProps<typeof SvgIcon>;
+}
 
-const IconBusiness = React.forwardRef((props: Props, ref: React.Ref<SVGSVGElement>) => (
-  <IconBusinessComponent {...props} forwardRef={ref} />
-));
+const IconBusiness = React.forwardRef<SVGSVGElement, SVGIconProps>(
+  (props: SVGIconProps, ref: React.Ref<SVGSVGElement>) => (
+    <IconBusinessComponent {...props} forwardRef={ref} />
+  )
+);
 
 IconBusiness.displayName = 'IconBusiness';