--- a/static/app/icons/iconCalendar.tsx
+++ b/static/app/icons/iconCalendar.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 
-import SvgIcon from './svgIcon';
+import SvgIcon, {SVGIconProps} from './svgIcon';
 
-type Props = React.ComponentProps<typeof SvgIcon>;
-
-const IconCalendar = React.forwardRef(function IconCalendar(
-  props: Props,
+const IconCalendar = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconCalendar(
+  props: SVGIconProps,
   ref: React.Ref<SVGSVGElement>
 ) {
   return (