**This is a draft PR that will not get merged, it is posted just as a proof of concept**

TLDR: we can cut TSC time almost in half and have a nicer and more responsive development experience

Longer story:
I wanted to understand why the typescript development flow on my machine was slow - code actions on save would take several seconds before type checking was done and errors due to incorrect types would pop-up, this was problematic because it decreased my development speed and resulted in multiple back and fourths between CI and TSC failing to build.

I've written down the investigation process and findings in https://www.notion.so/sentry/Typescript-performance-audit-ac375cf3d9374e66998a6eab2d5539e4

Before ~36s build time:
<img width="1680" alt="before" src="https://user-images.githubusercontent.com/9317857/147479310-f10311ef-8be6-47ea-8c6c-a447d5970e0c.png">
After ~21s
<img width="1680" alt="after" src="https://user-images.githubusercontent.com/9317857/147479346-2e52ea25-7634-4320-b986-6c05d6a3f066.png">

--diagnostics result before:
```
Files:              4008
Lines:            644694
Nodes:           2473488
Identifiers:      803265
Symbols:         3235184
Types:           1003633
Instantiations:  3280536
Memory used:    2772695K
I/O read:          0.46s
I/O write:         0.00s
Parse time:        2.93s
Bind time:         1.16s
Check time:       31.08s
Emit time:         0.00s
Total time:       35.16s
```
--diagnostics result after
```
Files:              4010 (+2)
Lines:            644522 (-172)
Nodes:           2468186 (-5302)
Identifiers:      801193 (-2072)
Symbols:         1627093 (-1 608 091)
Types:            540442 (-463 191)
Instantiations:  2499616 (-780 920
Memory used:    1725751K (-1 046 944K ~ 1GB)
I/O read:          0.43s (-0.03s)
I/O write:         0.00s (no change)
Parse time:        2.78s (-0.15s)
Bind time:         1.20s (+0.04s)
Check time:       17.15s (-13.93s)
Emit time:         0.00s (no change)
Total time:       21.13s (-14.03s)
```

The symbols, types, instantiations and memory usage are probably the most notable drops.