@evanpurkhiser 2021-12-28
Can't see the notion document right now, but wondering if we could split this out into a few different PRs for sets of changes?