@JonasBa 2021-12-28
Ah yes, sorry it was still private 🤦🏼‍♂️ - https://www.notion.so/sentry/Typescript-performance-audit-ac375cf3d9374e66998a6eab2d5539e4 I gave access to engineering